move icons file to components folder
